Fixed some bugs on checkList.
