+++ app/src/main/java/org/secfirst/umbrella/fragments/TabbedFragment.java
@@ -1,8 +1,6 @@
 package org.secfirst.umbrella.fragments;
 
-import android.app.AlertDialog;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.res.Configuration;
 import android.net.Uri;
@@ -14,19 +12,12 @@ import android.support.v4.view.ViewPager;
 import android.support.v7.widget.CardView;
 import android.view.Gravity;
 import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
-import android.widget.EditText;
 import android.widget.GridView;
-import android.widget.ImageButton;
-import android.widget.ListView;
-import android.widget.ProgressBar;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.j256.ormlite.stmt.QueryBuilder;
 import com.j256.ormlite.stmt.Where;
@@ -34,14 +25,9 @@ import com.j256.ormlite.stmt.Where;
 import org.secfirst.umbrella.AboutActivity;
 import org.secfirst.umbrella.MainActivity;
 import org.secfirst.umbrella.R;
-import org.secfirst.umbrella.adapters.CheckListAdapter;
 import org.secfirst.umbrella.adapters.GridAdapter;
-import org.secfirst.umbrella.models.CheckItem;
-import org.secfirst.umbrella.models.Difficulty;
-import org.secfirst.umbrella.models.Favourite;
 import org.secfirst.umbrella.models.Segment;
 import org.secfirst.umbrella.util.Global;
-import org.secfirst.umbrella.util.UmbrellaUtil;
 
 import java.sql.SQLException;
 import java.util.List;
@@ -90,7 +76,7 @@ public class TabbedFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         View v = inflater.inflate(R.layout.fragment_tabbed, container, false);
         mSectionsPagerAdapter = new SectionsPagerAdapter(getChildFragmentManager(), difficulty);
-        mViewPager = (ViewPager) v.findViewById(R.id.pager);
+        mViewPager = v.findViewById(R.id.pager);
         mViewPager.setAdapter(mSectionsPagerAdapter);
         mViewPager.setOnPageChangeListener(new ViewPager.OnPageChangeListener() {
             @Override
@@ -196,7 +182,7 @@ public class TabbedFragment extends Fragment {
             if (position == 0) {
                 fragment = new TabbedContentFragment();
             } else if (position == segments.size() + 1) {
-                fragment = new CheckItemFragment();
+                fragment = CheckItemFragment.newInstance(difficulty);
             } else {
                 fragment = new TabbedSegmentFragment();
                 args.putInt(TabbedFragment.ARG_SEGMENT_INDEX, position - 1);
@@ -248,15 +234,15 @@ public class TabbedFragment extends Fragment {
                         .and().like(Segment.FIELD_DIFFICULTY, difficulty);
                 final List<Segment> segments = queryBuilder.query();
                 if (!segments.isEmpty()) {
-                    GridView gridView = (GridView) rootView.findViewById(R.id.grid_tiles);
+                    GridView gridView = rootView.findViewById(R.id.grid_tiles);
                     GridAdapter gAdapter = new GridAdapter(getActivity(), segments);
                     gridView.setAdapter(gAdapter);
                 }
-                TextView toChecklist = (TextView) rootView.findViewById(R.id.grid_title);
+                TextView toChecklist = rootView.findViewById(R.id.grid_title);
                 toChecklist.setText(getString(R.string.checklist));
                 int[] colours = {R.color.umbrella_purple, R.color.umbrella_green, R.color.umbrella_yellow};
                 toChecklist.setBackgroundColor(getActivity().getResources().getColor(colours[(segments.size()) % 3]));
-                CardView checklistCard = (CardView) rootView.findViewById(R.id.checklist_view);
+                CardView checklistCard = rootView.findViewById(R.id.checklist_view);
                 if (drawerItem == 56) checklistCard.setVisibility(View.GONE);
                 checklistCard.setOnClickListener(new View.OnClickListener() {
                     @Override
@@ -321,7 +307,7 @@ public class TabbedFragment extends Fragment {
                 public void onLoadResource(WebView view, String url) {
                 }
             };
-            content = (WebView) rootView.findViewById(R.id.segment_content);
+            content = rootView.findViewById(R.id.segment_content);
             content.setWebViewClient(interceptUrlLicences);
 
             int drawerItem = (int) ((MainActivity) getActivity()).drawerItem;
@@ -351,203 +337,4 @@ public class TabbedFragment extends Fragment {
         }
 
     }
-
-    public static class CheckItemFragment extends Fragment {
-
-        private List<CheckItem> mCheckList;
-        private ProgressBar checkBar;
-        private TextView checkBarText;
-        private CheckListAdapter cLAdapter;
-
-        public CheckItemFragment() {
-        }
-
-        @Override
-        public void onCreate(Bundle savedInstanceState) {
-            super.onCreate(savedInstanceState);
-            setHasOptionsMenu(true);
-        }
-
-        @Override
-        public void onAttach(Context context) {
-            if (!context.getResources().getConfiguration().locale.toString().equals(Locale.getDefault().toString())) {
-                Configuration config = new Configuration();
-                config.locale = Locale.getDefault();
-                context.getResources().updateConfiguration(config, context.getResources().getDisplayMetrics());
-            }
-            super.onAttach(context);
-        }
-
-        @Override
-        public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                                 Bundle savedInstanceState) {
-            View rootView = inflater.inflate(R.layout.fragment_check_item,
-                    container, false);
-
-            final long drawerItem = ((MainActivity) getActivity()).drawerItem;
-            ListView contentBox = rootView.findViewById(R.id.content_box);
-            checkBar = rootView.findViewById(R.id.progress_bar_checked);
-            checkBarText = rootView.findViewById(R.id.check_bar_text);
-            setProgressBarTo(0);
-            final int diffArg = getArguments().getInt(ARG_DIFFICULTY_NUMBER, 1);
-            ImageButton addItem = rootView.findViewById(R.id.fab);
-            addItem.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    if (!Global.INSTANCE.hasPasswordSet(false)) {
-                        Global.INSTANCE.setPassword(getActivity(), null);
-                    } else {
-                        AlertDialog.Builder alert = new AlertDialog.Builder(getActivity());
-                        alert.setTitle(R.string.add_new_checkitem);
-                        alert.setMessage(R.string.add_own_checklist_item);
-                        final EditText pwInput = new EditText(getActivity());
-                        alert.setView(pwInput);
-                        alert.setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
-                            public void onClick(DialogInterface dialog, int whichButton) {
-                                String pw = pwInput.getText().toString();
-                                if (pw.length() > 4) {
-                                    CheckItem nItem = new CheckItem(pw, (int) drawerItem);
-                                    nItem.setCustom(1);
-                                    nItem.setDifficulty(diffArg);
-                                    nItem.setDifficultyString(UmbrellaUtil.a(diffArg));
-                                    try {
-                                        Global.INSTANCE.getDaoCheckItem().create(nItem);
-                                    } catch (SQLException e) {
-                                        Timber.e(e);
-                                    }
-                                    refreshCheckList(drawerItem, diffArg);
-                                    dialog.dismiss();
-                                    Toast.makeText(getActivity(), R.string.you_have_added_new_item, Toast.LENGTH_SHORT).show();
-                                } else {
-                                    Toast.makeText(getActivity(), R.string.text_item_has_to_be_longer, Toast.LENGTH_SHORT).show();
-                                }
-                            }
-                        });
-                        alert.setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
-                            public void onClick(DialogInterface dialog, int whichButton) {
-                                dialog.cancel();
-                            }
-                        });
-                        alert.show();
-                    }
-                }
-            });
-
-            refreshCheckList(drawerItem, diffArg);
-            cLAdapter = new CheckListAdapter(getActivity(), mCheckList, this);
-            contentBox.setAdapter(cLAdapter);
-            contentBox.setDivider(null);
-            return rootView;
-        }
-
-        @Override
-        public void setMenuVisibility(boolean menuVisible) {
-            super.setMenuVisibility(menuVisible);
-            if (menuVisible) {
-                if (getActivity() != null) setFavouriteIcon(getActivity());
-            } else {
-                if (getActivity() != null)
-                    ((MainActivity) getActivity()).favouriteItem.setVisible(false);
-            }
-        }
-
-        @Override
-        public void onPrepareOptionsMenu(Menu menu) {
-            super.onPrepareOptionsMenu(menu);
-            if (getActivity() != null) setFavouriteIcon(getActivity());
-        }
-
-        @Override
-        public boolean onOptionsItemSelected(MenuItem item) {
-            int id = item.getItemId();
-            if (getActivity() != null) {
-                final long drawerItem = ((MainActivity) getActivity()).drawerItem;
-                if (id == R.id.favourite) {
-                    List<Difficulty> hasDifficulty = null;
-                    try {
-                        hasDifficulty = Global.INSTANCE.getDaoDifficulty().queryForEq(Difficulty.FIELD_CATEGORY, String.valueOf(drawerItem));
-                    } catch (SQLException e) {
-                        Timber.e(e);
-                    }
-                    if (hasDifficulty != null && !hasDifficulty.isEmpty()) {
-                        try {
-                            QueryBuilder<Favourite, String> queryBuilder = Global.INSTANCE.getDaoFavourite().queryBuilder();
-                            Where<Favourite, String> where = queryBuilder.where();
-                            where.eq(Favourite.FIELD_CATEGORY, String.valueOf(drawerItem)).and().eq(Favourite.FIELD_DIFFICULTY, String.valueOf(hasDifficulty.get(0).getSelected()));
-                            Favourite favourite = queryBuilder.queryForFirst();
-                            if (favourite != null) {
-                                try {
-                                    Global.INSTANCE.getDaoFavourite().delete(favourite);
-                                } catch (SQLException e) {
-                                    Timber.e(e);
-                                }
-                            } else {
-                                try {
-                                    Global.INSTANCE.getDaoFavourite().create(new Favourite(drawerItem, hasDifficulty.get(0).getSelected()));
-                                } catch (SQLException e) {
-                                    Timber.e(e);
-                                }
-                            }
-                        } catch (SQLException e) {
-                            Timber.e(e);
-                        }
-                    }
-                    setFavouriteIcon(getActivity());
-                    getActivity().invalidateOptionsMenu();
-                    return true;
-                }
-            }
-            return super.onOptionsItemSelected(item);
-        }
-
-        public void setFavouriteIcon(Context context) {
-            final long drawerItem = ((MainActivity) context).drawerItem;
-            try {
-                QueryBuilder<Favourite, String> queryBuilder = Global.INSTANCE.getDaoFavourite().queryBuilder();
-                Where<Favourite, String> where = queryBuilder.where();
-                where.eq(Favourite.FIELD_CATEGORY, String.valueOf(drawerItem)).and().eq(Favourite.FIELD_DIFFICULTY, String.valueOf(difficulty));
-                Favourite favourite = queryBuilder.queryForFirst();
-                ((MainActivity) context).favouriteItem.setIcon(favourite != null ? R.drawable.abc_btn_rating_star_on_mtrl_alpha : R.drawable.abc_btn_rating_star_off_mtrl_alpha);
-                ((MainActivity) context).favouriteItem.setVisible(true);
-            } catch (SQLException e) {
-                Timber.e(e);
-            }
-        }
-
-        public void refreshCheckList(long category, int difficulty) {
-            try {
-                QueryBuilder<CheckItem, String> queryBuilder = Global.INSTANCE.getDaoCheckItem().queryBuilder();
-                Where<CheckItem, String> where = queryBuilder.where();
-                where.eq(CheckItem.FIELD_CATEGORY, String.valueOf(category)).and().eq(CheckItem.FIELD_DIFFICULTY, String.valueOf(difficulty));
-                mCheckList = queryBuilder.query();
-            } catch (SQLException e) {
-                Timber.e(e);
-            }
-
-            if (mCheckList != null) {
-                if (cLAdapter != null) {
-                    cLAdapter.updateData(mCheckList);
-                }
-                if (!mCheckList.isEmpty()) {
-                    int selected = 0;
-                    int total = 0;
-                    for (CheckItem checkItem : mCheckList) {
-                        if (!checkItem.getNoCheck() && !checkItem.isDisabled()) {
-                            total++;
-                            if (checkItem.getValue()) selected++;
-                        }
-                    }
-                    setProgressBarTo((int) Math.round(selected * 100.0 / total));
-                }
-            }
-        }
-
-        public void setProgressBarTo(int percent) {
-            if (percent >= 0 && percent <= 100) {
-                checkBar.setProgress(percent);
-                checkBarText.setText(percent + "% " + getActivity().getString(R.string.filled));
-            }
-        }
-    }
-
 }
