+++ app/src/main/res/layout/fragment_check_item.xml
@@ -1,22 +1,22 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:orientation="vertical"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
     <LinearLayout
-        android:paddingLeft="5dp"
-        android:paddingRight="5dp"
         android:layout_width="fill_parent"
         android:layout_height="match_parent"
-        android:orientation="vertical">
+        android:orientation="vertical"
+        android:paddingLeft="5dp"
+        android:paddingRight="5dp">
 
         <LinearLayout
             android:id="@+id/progress_checked"
-            android:orientation="vertical"
-            android:layout_marginBottom="5dp"
             android:layout_width="fill_parent"
-            android:layout_height="wrap_content">
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="5dp"
+            android:orientation="vertical">
 
             <ProgressBar
                 android:id="@+id/progress_bar_checked"
@@ -26,20 +26,20 @@
 
             <TextView
                 android:id="@+id/check_bar_text"
+                android:layout_width="fill_parent"
+                android:layout_height="wrap_content"
                 android:gravity="center"
                 android:textColor="@color/umbrella_green"
-                android:textSize="@dimen/text_size_small"
-                android:layout_width="fill_parent"
-                android:layout_height="wrap_content" />
+                android:textSize="@dimen/text_size_small" />
 
         </LinearLayout>
 
-        <ListView
+        <android.support.v7.widget.RecyclerView
             android:id="@+id/content_box"
-            android:listSelector="@android:color/transparent"
-            android:cacheColorHint="@android:color/transparent"
             android:layout_width="fill_parent"
-            android:layout_height="fill_parent" />
+            android:layout_height="fill_parent"
+            android:cacheColorHint="@android:color/transparent"
+            android:listSelector="@android:color/transparent" />
 
     </LinearLayout>
 
@@ -51,18 +51,17 @@
 
         <ImageButton
             android:id="@+id/fab"
-            android:layout_alignParentRight="true"
-            android:layout_alignParentBottom="true"
             android:layout_width="@dimen/fab_size"
             android:layout_height="@dimen/fab_size"
+            android:layout_alignParentBottom="true"
+            android:layout_alignParentRight="true"
             android:layout_gravity="bottom|right"
-            android:layout_marginRight="16dp"
             android:layout_marginBottom="16dp"
+            android:layout_marginRight="16dp"
             android:background="@drawable/round_button"
-            android:stateListAnimator="@animator/anim"
-            android:src="@drawable/ic_add"
             android:elevation="4dp"
-            />
+            android:src="@drawable/ic_add"
+            android:stateListAnimator="@animator/anim" />
 
     </RelativeLayout>
